[DistDGL] back-compatible with dgl.dataloading
